adding randomizing

+++ sources/order.cpp
@@ -3,7 +3,7 @@
 #include "../headers/Order.h"
 
 Order::Order(Order::Order_Type type, Status status,
-             std::uint64_t order_ID, float price, int quantity, int filled_quantity)
+             std::uint64_t order_ID, double price, int quantity, int filled_quantity)
     : type{type}, status{status}, order_ID{order_ID}, price{price}, quantity{quantity}, filled_quantity{0} {}
 
 void Order::cancel() {
